improvement(user): extracts composeUser function from handleIntegration for easier testing
